Extract IPackageManager interface and reorganize package management code
Move package management files to a dedicated packages/ subdirectory and extract
IPackageManager interface to decouple session from concrete implementation.

See #11662
